Add files via upload
Classification is a supervised learning approach in which the computer program learns from the data input given to it and then uses this learning to classify new observation. This data set may simply be bi-class (like identifying whether the person is male or female or that the mail is spam or non-spam) or it may be multi-class too. Some examples of classification problems are: speech recognition, handwriting recognition, bio metric identification, document classification etc.The Repository Contains Source code in Python as well as R-Programming. This Repository contains Source code of 7 Different Classification  Techniques. 
1) Logistic Regression 
2) K-Nearest Neighbour
3) Support Vector Machine
4) Kernel SVM
5) Naive Bayes
6) Decision Tree Regression
7) Random forest Regression 
Refer to the Dataset Repository to get the dataset for all the models and algorithms.